refactor(infer): deduplicate literal_provides_optional_class_field
Move the function to narrow/mod.rs (pub(in crate::semantic)),
re-export it via infer/mod.rs, and remove the identical copy
from both get_type_at_flow.rs and semantic_info/mod.rs.

Also fix two issues in write_instance_type (humanize_type.rs):
- compute without_nil once instead of three times in the same loop
- guard function_vec loop with count < max_display_count instead
  of count < all_count
