+++ crates/emmylua_code_analysis/src/semantic/infer/mod.rs
@@ -27,6 +27,7 @@ use infer_table::infer_table_expr;
 pub use infer_table::{infer_table_field_value_should_be, infer_table_should_be};
 use infer_unary::infer_unary_expr;
 pub(in crate::semantic) use narrow::ConditionFlowAction;
+pub(in crate::semantic) use narrow::literal_provides_optional_class_field;
 pub use narrow::VarRefId;
 
 use rowan::TextRange;
